Move date utils into redshift handler since it only used there.
